moving the bigger components into a new folder sections
